completed openAI integration

+++ src/app/(routes)/dashboard/incomes/_components/IncomeList.jsx
@@ -1,9 +1,9 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import CreateIncomes from "./CreateIncomes";
-import { db } from "@/utils/dbConfig";
+import { db } from "../../../../../../utils/dbConfig";
 import { desc, eq, getTableColumns, sql } from "drizzle-orm";
-import { Incomes, Expenses } from "@/utils/schema";
+import { Incomes, Expenses } from "../../../../../../utils/schema";
 import { useUser } from "@clerk/nextjs";
 import IncomeItem from "./IncomeItem";
 
@@ -18,7 +18,7 @@ function IncomeList() {
     const result = await db
       .select({
         ...getTableColumns(Incomes),
-        totalSpend: sql`sum(${Expenses.amount})`.mapWith(Number),
+        totalSpend: sql`sum((${Expenses.amount})::numeric)`.mapWith(Number),
         totalItem: sql`count(${Expenses.id})`.mapWith(Number),
       })
       .from(Incomes)
